finished dbclasses

+++ domain/src/main/java/ru/javacat/domain/models/Customer.kt
@@ -2,7 +2,7 @@ package ru.javacat.domain.models
 
 data class Customer (
     val companyName: String,
-    val atiNumber: Long,
+    val atiNumber: Int,
     val employeeList: List<Employee>?,
     val companyPhone: Long?,
 )
